+++ CHANGELOG.md
@@ -28,6 +28,7 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
   which enables more efficient iteration. As a result `.coco_images_iter` is deprecated.
 * Modify categories keep argument will coerce to a list if given as a string.
 * Confusion measures will now included more metrics at maximized operating points.
+* Segmentation metrics now respect truth annotation weights.
 
 ### Fixed
 * index failure if tracks were used on images without `frame_index`